requirements.txt
certifi==2021.5.30
Cython==0.29.23
joblib==1.0.1
mkl-fft>=1.3.0
mkl-random
mkl-service
numpy
olefile==0.46
pandas==1.2.4
Pillow
python-dateutil==2.8.1
pytz==2021.1
quadprog
scikit-learn
scipy
six
threadpoolctl==2.1.0
#torch==1.8.1
torchaudio
torchvision
tqdm






